change option from doc to index
