+++ doc.js
@@ -1,19 +1,19 @@
-const options = {
-  definition: {
-    openapi: "3.1.0",
-    info: {
-      title: "Dota2 Heros API",
-      version: "0.1.0",
-      description: "Dota 2 heros api by Florject",
-    },
-    servers: [
-      {
-        url: "https://abusing-scripts.vercel.app",
-        // url: "http://localhost:3000",
-      },
-    ],
-  },
-  apis: [ __dirname+"/*.js" ],
-};
+// const options = {
+//   definition: {
+//     openapi: "3.1.0",
+//     info: {
+//       title: "Dota2 Heros API",
+//       version: "0.1.0",
+//       description: "Dota 2 heros api by Florject",
+//     },
+//     servers: [
+//       {
+//         url: "https://abusing-scripts.vercel.app",
+//         // url: "http://localhost:3000",
+//       },
+//     ],
+//   },
+//   apis: [ __dirname+"/*.js" ],
+// };
 
-module.exports = options;
+// module.exports = options;
